requirements.txt
pandas
openpyxl
requests
selenium
webdriver-manager
python-dotenv
behave
rich
